requirements.txt
aiogram>=3.20.0
aiogram-tonconnect>=0.15.0
asyncpg>=0.30.0
fluentogram>=1.2.0
pydantic>=2.11.0
pytonconnect>=0.2.0
redis>=6.2.0
solders>=0.26.0
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
babel
base58
backoff
certifi
charset-normalizer
construct
construct-typing
fluent-compiler
fluent.syntax
frozenlist
h11
httpcore
httpx
idna
jsonalias
magic-filter
multidict
propcache
python-dotenv
pytz
PyYAML
requests
sniffio
solana
typing-inspection
typing_extensions
tonsdk
urllib3
websockets
yarl
